backend/users/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.8
python-dotenv
hvac
djangorestframework
certifi
requests
gunicorn
django-phonenumber-field
django-phonenumbers
Pillow
djangorestframework-simplejwt
